+++ packages/ice/src/plugins/web/index.ts
@@ -1,4 +1,5 @@
 import * as path from 'path';
+import fse from 'fs-extra';
 import chalk from 'chalk';
 import type { RenderMode } from '@ice/runtime';
 import lodash from '@ice/bundles/compiled/lodash/index.js';
@@ -6,13 +7,15 @@ import type { Plugin } from '../../types/plugin.js';
 import ReCompilePlugin from '../../webpack/ReCompilePlugin.js';
 import DataLoaderPlugin from '../../webpack/DataLoaderPlugin.js';
 import { getRouteExportConfig } from '../../service/config.js';
-import { WEB, SERVER_OUTPUT_DIR } from '../../constant.js';
+import { WEB, SERVER_OUTPUT_DIR, IMPORT_META_TARGET, IMPORT_META_RENDERER } from '../../constant.js';
 import getWebTask from '../../tasks/web/index.js';
 import generateHTML from '../../utils/generateHTML.js';
 import openBrowser from '../../utils/openBrowser.js';
 import getServerCompilerPlugin from '../../utils/getServerCompilerPlugin.js';
 import type ServerCompilerPlugin from '../../webpack/ServerCompilerPlugin.js';
 import { logger } from '../../utils/logger.js';
+import getRoutePaths from '../../utils/getRoutePaths.js';
+import getRouterManifest from '../../utils/getRouterManifest.js';
 
 const { debounce } = lodash;
 
@@ -67,6 +70,10 @@ const plugin: Plugin = () => ({
         serverCompileTask: command === 'start' ? serverCompileTask : null,
         userConfig,
         ensureRoutesConfig,
+        runtimeDefineVars: {
+          [IMPORT_META_TARGET]: JSON.stringify('web'),
+          [IMPORT_META_RENDERER]: JSON.stringify('server'),
+        },
         incremental: command === 'start',
       });
       webpackConfigs[0].plugins.push(
@@ -89,10 +96,8 @@ const plugin: Plugin = () => ({
           }),
         );
         const debounceCompile = debounce(() => {
+          serverCompilerPlugin?.buildResult?.rebuild();
           console.log('Document updated, try to reload page for latest html content.');
-          if (serverCompilerPlugin) {
-            serverCompilerPlugin.compileTask();
-          }
         }, 200);
         watch.addEvent([
           /src\/document(\/index)?(.js|.jsx|.tsx)/,
@@ -112,7 +117,7 @@ const plugin: Plugin = () => ({
         renderMode = 'SSG';
       }
       serverEntryRef.current = serverOutfile;
-
+      const routeType = appConfig?.router?.type;
       await generateHTML({
         rootDir,
         outputDir,
@@ -122,6 +127,24 @@ const plugin: Plugin = () => ({
         renderMode,
         routeType: appConfig?.router?.type,
       });
+
+      if (routeType === 'memory' && userConfig?.routes?.injectInitialEntry) {
+        // Read the latest routes info.
+        const routes = getRouterManifest(rootDir);
+        const routePaths = getRoutePaths(routes);
+        routePaths.forEach((routePath) => {
+          // Inject `initialPath` when router type is memory.
+          const routeAssetPath = path.join(outputDir, 'js',
+            `p_${routePath === '/' ? 'index' : routePath.replace(/^\//, '').replace(/\//g, '-')}.js`);
+          if (fse.existsSync(routeAssetPath)) {
+            fse.writeFileSync(routeAssetPath,
+              `window.__ICE_APP_CONTEXT__=Object.assign(window.__ICE_APP_CONTEXT__||{}, {routePath: '${routePath}'});${
+              fse.readFileSync(routeAssetPath, 'utf-8')}`);
+          } else {
+            logger.warn(`Can not find ${routeAssetPath} when inject initial path.`);
+          }
+        });
+      }
     });
 
     onHook('after.start.compile', async ({ isSuccessful, isFirstCompile, urls, devUrlInfo }) => {