+++ packages/plugin-pha/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - [feat] support resource_prefetch to preload resource
 - [feat] support enableExpiredManifest
+- [feat] support configure `enableExpiredManifest`
 
 ## 1.1.0